[Refactor] 루트 레이아웃 구조 변경
